+++ src/components/pages/community/PostEditPage.tsx
@@ -15,7 +15,7 @@ import { fetchPostById, updatePost, deletePostById } from './api/postApi';
 import CategoryWrapper from '../../common/CategoryWrapper';
 import { POST_CATEGORIES } from './postCategories';
 import { Post, PostDetailResponse } from '../../../types/postTypes';
-import { getImageSrc } from '../../../hooks/GetImageSrc';
+import { getImageSrc } from '../../../utils/GetImageSrc';
 
 const PostEditPage = () => {
   const { communityPostId } = useParams<{ communityPostId: string }>();
